Fix publish target
I have converted the repository from make to mage, and from azure devops
pipelines to github actions. I figured as long as I was here fixing our
broken build, lets just finish moving over to where we want to be.

The problem with the publish target was a few things. One there was a
bug in the magefiles generatePackageFeed function, where it wasn't
checking the error when calling porter to generate the atom.xml file for
the feed. Porter wasn't installed, and it was quietly failing.

We are now using a fixed version of the magefiles repo, and installing
porter into bin before attempting to publish.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>